runtime: Handle auxv data for ppc64 and ppc64le
This commit saves the values of AT_HWCAP and AT_HWCAP2 from the auxiliary
vector, so they can be used later for runtime checking the ISA level and
hardware capabilities to enable some instructions. It also defines a few
hwcap/hwcap2 masks that will be commonly used for such runtime checks, and also
provides a function for these verifications.

These checks will be used more often as new POWER8/POWER9 optimizations are
added to the ppc64x runtime, so they do not break compatibility with older
processors.